feat: Integrating kv value server with the KeyFetcherManagerInterface
The interface is reponsible for getting/caching/refreshing the public and private keys and mappings to key ids.
In this CL we're actually constucting a fake KeyFetcherManager, that always returns hardcoded public/private key values and ids. The follow up CL will construct the correct version. The purpose of this CL is to plumb this logic through.
Note that we're changing the test keys to match the values of what we have in the common repo.

E2E tested with https://source.corp.google.com/h/team/kiwi-air-force-eng-team/kv-server/+/main:docs/testing_the_query_protocol.md
that had was looking up keys on two different data servers. So the client to server was encrypted, and so was UDF -> data server.

Bug: 290796192
Change-Id: I9fc856cdb2cd53ae2f0d6ed32278c4107936b989
GitOrigin-RevId: 4acb488a9aaf9e3fd0f53d3391f29f34a1c308d9