chore: Add coordinator support to query validator.
This change added coordinator support to the query validator by using key_fetcher_manager and the ObliviousGetValues rpc. All variables (use_real_coordinator flag, and all the coordinator related variables) required by the key_fetcher_manager are acquired with the help of a parameter_fetcher. And these variables are ultimately from terraform variable files we set for each environment. Since the same set of variables are also used by the actual kv-server for that environment, the e2e test validator setup should be inline with the kv-server under test.

Note that we also have a separate variable (e2e test variable) called use_coordinator. And  the validator would only use the aforementioned key_fetcher_manager if the e2e variable use_coordinator is set to true. If use_coordinator is set to false, we won’t use coordinator and ObliviousGetValues for query validation regardless of terraform variable setup.

This change is tested (manually triggered e2e) with use_coordinator
setting to both true and false for both AWS and GCP.

Bug: b/335279300
Change-Id: I532880e27203f2d09a296baf3653ed209aadd453
GitOrigin-RevId: f85baee81fd6428dd7db29c2b2433f01ce0463a1