chore: Temporarily remove PS_VLOG from GCP instance client until permanent fix is implemented
Logging verbosity will not be able to set properly if PS_VLOG is called before
verbosity parameter is retrieved and set. This is something that is fundamentally wrong in the current
PS_VLOG_IS_ON implemention in the common repo and should be fixed from there.
Details see in the buganizer

Bug: b/342614468
Change-Id: Id6e343c03802a4389d2d07a02b91b1bc091340b9
GitOrigin-RevId: 501abd3e4a4a443120eb3c9356092795fc4bc936